Major Update: Additional Repo
As Bounty Hunters we use many tools, some which are not available on every distro as a Hacker's distro Kali is guaranteed to have most tools if not all. This pull request addresses the issue of going far and wide to install tools which are not prepackaged in normal distro's by directly adding kali repo's. The repo priority is low so as not to clog the system with it's own bolahoo packages {close to 1.5K packages}. Tools can be installed by the use of aptitude which handles dependency issues better than apt, apt-get or apt-fast. WPScan is a great example of a Kali packaged tool. I have tested and verified that this addition does not mess with the workings of this tool. :smiley: 

PS: A .DS_STORE file parser has also been added (very handy) but the new repo is the main attraction for now.

Thanks,
CowlingBanana